+++ reductstore/src/lib.rs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0
 pub mod api;
 pub mod asset;
-pub mod audit;
 pub mod auth;
 pub mod backend;
 pub mod cfg;
@@ -13,3 +12,4 @@ pub mod lifecycle;
 pub mod lock_file;
 pub mod replication;
 pub mod storage;
+pub mod syslog;